requirements.txt
python-dotenv
yagmail
requests
beautifulsoup4
google-genai
